TA allocation page integrated with backend

+++ app/Spring-Microservices/applicationservice/src/main/java/com/infinity/applicationservice/dtos/Allocations/AllocationRequest.java
@@ -9,6 +9,6 @@ public record AllocationRequest(
         Long studentId,
         Long applicationId,
         TaskType task,
-        int hours,
+        double hours,
         Long sectionId) {
 }
